World entities are now stored in a separate file each instead of all being saved inside a single hwld file This makes collaboration with multiple content creatores easier

+++ src/mmo_edit/editors/world_editor/edit_modes/world_edit_mode.h
@@ -43,7 +43,7 @@ namespace mmo
 
 		virtual float GetTranslateGridSnapSize() const = 0;
 
-		virtual Entity* CreateMapEntity(const String& assetName, const Vector3& position, const Quaternion& orientation, const Vector3& scale, uint32 objectId) = 0;
+		virtual Entity* CreateMapEntity(const String& assetName, const Vector3& position, const Quaternion& orientation, const Vector3& scale, uint64 objectId) = 0;
 
 		virtual bool HasTerrain() const = 0;
